+++ NEWS.md
@@ -13,3 +13,8 @@
 # spotifyr 2.2.0
 * Release candidate for CRAN.
 * Resubmitted after reformatting two (possibly) invalid urls.
+
+# spotifyr 2.2.1
+* All functions in the documentation have well-specified return values.
+* Assertions are made with assertthat for more meaningful error messages.
+* Release candidate for CRAN.